Extract shared offline test harness into backend/tests/harness.py
